% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_rules}
\alias{simplify_rules}
\title{Simplify rules to the minimal defining set}
\usage{
simplify_rules(rules)
}
\arguments{
\item{rules}{A vector of rules or a list of rule components.}
}
\value{
A vector or a list, based on the input, with the smallest set of
rules that identifies a partition.
}
\description{
Tree algorithms returns nested, redundant rule sets. This function simplify
the rules keeping the shortest condition set required to univocally identify
the partition.
}
\examples{

tree <- ctree(Sepal.Length ~ Species + Sepal.Width, iris)
rules <- tidy_tree()
simplify_rules(rules$rule)

## Works also with a list of conditions

#' rules <- tidy_tree(rule_as_text = FALSE)
simplify_rules(rules$rule)

}
